+++ src/geometry.hpp
@@ -10,18 +10,18 @@
 #include "helper.hpp"
 #include "material.hpp"
 
-const float EPSILON = 0.0000001f;
-const float OFFSET = 0.0000001f;
-const float PADDING = 0.01f;
+const double EPSILON = 0.0000001f;
+const double OFFSET = 0.0000001f;
+const double PADDING = 0.01f;
 
 struct BoundingBox
 {
-    glm::vec3 vMin, vMax;
+    glm::dvec3 vMin, vMax;
 
-    BoundingBox() : vMin(std::numeric_limits<float>::max()), vMax(std::numeric_limits<float>::min()) {}
+    BoundingBox() : vMin(std::numeric_limits<double>::max()), vMax(std::numeric_limits<double>::min()) {}
 
-    glm::vec3 centroid() { return 0.5f * vMin + 0.5f * vMax; }
-    void merge(const glm::vec3 &other)
+    glm::dvec3 centroid() { return 0.5 * vMin + 0.5 * vMax; }
+    void merge(const glm::dvec3 &other)
     {
         vMin.x = std::min(vMin.x, other.x);
         vMin.y = std::min(vMin.y, other.y);
@@ -42,7 +42,7 @@ struct BoundingBox
 
     int maxAxis()
     {
-        glm::vec3 minus = vMax - vMin;
+        glm::dvec3 minus = vMax - vMin;
         if (minus.x > minus.y && minus.x > minus.z)
             return 0;
         else if (minus.y > minus.z)
@@ -53,13 +53,13 @@ struct BoundingBox
 
     bool intersect(const Ray &ray) const
     {
-        float tmin = 0.00000f;
-        float tmax = std::numeric_limits<float>().max();
-        glm::vec3 inv = 1.f / ray.direction;
+        double tmin = 0.00000f;
+        double tmax = std::numeric_limits<double>().max();
+        glm::dvec3 inv = 1. / ray.direction;
         for (int i = 0; i < 3; i++)
         {
-            float minTime = (vMin[i] - ray.origin[i]) * inv[i];
-            float maxTime = (vMax[i] - ray.origin[i]) * inv[i];
+            double minTime = (vMin[i] - ray.origin[i]) * inv[i];
+            double maxTime = (vMax[i] - ray.origin[i]) * inv[i];
             if (minTime > maxTime)
                 std::swap(minTime, maxTime);
 
@@ -81,7 +81,7 @@ inline BoundingBox unionBox(const BoundingBox &bb1, const BoundingBox &bb2)
     return ret;
 }
 
-inline BoundingBox unionBox(const BoundingBox &bb1, const glm::vec3 &other)
+inline BoundingBox unionBox(const BoundingBox &bb1, const glm::dvec3 &other)
 {
     BoundingBox ret;
     ret.vMin = glm::min(bb1.vMin, other);
@@ -94,27 +94,27 @@ class Geometry
 public:
     virtual BoundingBox getBoundingBox() = 0;
     virtual HitResult getIntersection(const Ray &ray) = 0;
-    virtual float getArea() = 0;
-    virtual void sample(HitResult &point, float &pdf) = 0;
+    virtual double getArea() = 0;
+    virtual void sample(HitResult &point, double &pdf) = 0;
 };
 
 class Triangle : public Geometry
 {
 public:
-    glm::vec3 v0, v1, v2;
-    glm::vec3 edge1, edge2; // v1 - v0, v2 - v0;
-    glm::vec2 uv0, uv1, uv2;
-    glm::vec3 faceNormal;
-    float areaSize;
+    glm::dvec3 v0, v1, v2;
+    glm::dvec3 edge1, edge2; // v1 - v0, v2 - v0;
+    glm::dvec2 uv0, uv1, uv2;
+    glm::dvec3 faceNormal;
+    double areaSize;
     std::shared_ptr<Material> mat;
 
 public:
-    Triangle(glm::vec3 _v0, glm::vec3 _v1, glm::vec3 _v2);
+    Triangle(glm::dvec3 _v0, glm::dvec3 _v1, glm::dvec3 _v2);
 
-    void setUVs(glm::vec2 _uv0, glm::vec2 _uv1, glm::vec2 _uv2);
-    float getArea() override;
+    void setUVs(glm::dvec2 _uv0, glm::dvec2 _uv1, glm::dvec2 _uv2);
+    double getArea() override;
     BoundingBox getBoundingBox() override;
     HitResult getIntersection(const Ray &ray) override;
-    void sample(HitResult &point, float &pdf) override;
+    void sample(HitResult &point, double &pdf) override;
 };
 #endif // __GEOMETRY_H__
