+++ src/record.hpp
@@ -10,7 +10,7 @@ class Material;
 
 struct ScatteredResult
 {
-    glm::vec3 attenuation;
+    glm::dvec3 attenuation;
     bool skipPDF;
     Ray rayOut;
 };
@@ -20,9 +20,9 @@ struct HitResult
     bool isHit = false;
     double distance = std::numeric_limits<double>::max();
     std::shared_ptr<Material> mat;
-    float area = std::numeric_limits<float>::max();
-    glm::vec3 hitWorldNormal{};
-    glm::vec3 hitWorldPosition{};
+    double area = std::numeric_limits<double>::max();
+    glm::dvec3 hitWorldNormal{};
+    glm::dvec3 hitWorldPosition{};
     double hitTime = std::numeric_limits<double>::max();
     glm::vec2 hitUV{};
 };
